Edit GUI

+++ qpoint3d.h
@@ -13,9 +13,9 @@ public:
 //  void setX(int x_){this->setX(x_);}
 //  void setY(int y_){this->setY(y_);}
     void setZ(float z_){z=z_;}
-    int getX(){return this->x();}
-    int getY(){return this->y();}
-    int getZ(){return z;}
+    float getX(){return this->x();}
+    float getY(){return this->y();}
+    float getZ(){return z;}
 
 };
